screen-wake-lock/wakelock-onrelease.https.html

<!DOCTYPE html>
<link rel="help" href="https://w3c.github.io/screen-wake-lock/#the-onrelease-attribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
promise_test(async t => {
  await test_driver.bless("screen wake lock");
  const lock = await navigator.wakeLock.request("screen");
  const releaseEventPromise = new Promise(resolve => {
    lock.onrelease = resolve;
  });
  await lock.release();
  const ev = await releaseEventPromise;
  assert_class_string(ev, "Event", "release() must fire an Event object");
  assert_equals(ev.target, lock, "The event's target must be the lock that was acquired");
  assert_true(ev.isTrusted);
  assert_false(ev.bubbles);
  assert_false(ev.cancelable);
}, "Test onreleased event firing and attributes");

promise_test(async ()=>{
  await test_driver.bless("screen wake lock");
  const lock = await navigator.wakeLock.request("screen");
  let result = "";
  const releaseEventPromise = new Promise(resolve => {
    lock.onrelease = ()=>{
      result += "onrelease event fired";
      resolve();
    }
  });

  const releasePromise = lock.release().then(() => {
    result += " then release promise resolved";
  });

  await Promise.all([releaseEventPromise, releasePromise]);
  assert_equals(result, "onrelease event fired then release promise resolved");
}, "Ensure onreleased is called before WakeLockSentinel.release() resolves");

</script>
